code cleanup and fix some typos
